Tca 21.02.1 (#74)
* remove pwfile ref; usage cleanup
* update masters to be instance generic
* rename masters to more generic instance
* force prereqs apt update
* fix imagef settings
* add net-tools to apt prereqs

+++ bin/tdh-instance-init.sh
@@ -12,10 +12,10 @@ fi
 
 # -----------------------------------
 
-names="m01 m02 m03"
 prefix="$TDH_GCP_PREFIX"
 
-zone=
+names=
+zone="$GCP_DEFAULT_ZONE"
 mtype="$GCP_DEFAULT_MACHINETYPE"
 bootsize="$GCP_DEFAULT_BOOTSIZE"
 disksize="$GCP_DEFAULT_DISKSIZE"
@@ -28,15 +28,16 @@ gcpcompute="${tdh_path}/gcp-compute.sh"
 master_id="master-id_rsa.pub"
 master_id_file="${tdh_path}/../ansible/.ansible/${master_id}"
 
-dryrun=0
 attach=0
+disknum=1
+dryrun=0
 ssd=0
+xfs=0
 tags=
 action=
-rt=
 
-# -----------------------------------
-# default overrides
+# ----------------------------------
+# Default overrides
 
 if [ -n "$GCP_ZONE" ]; then
     zone="$GCP_ZONE"
@@ -57,29 +58,30 @@ fi
 # -----------------------------------
 
 usage="
-A script for creating TDH Master instances on GCP.
+A script for creating TDH instances on GCP.
 
 Synopsis:
   $TDH_PNAME [options] <action> host1 host2 ...
 
 Options:
-  -A|--attach           : Create an attached volume.
-  -b|--bootsize <xxGB>  : Size of boot disk, Default is $bootsize.
-  -d|--disksize <xxGB>  : Size of attached disk, Default is $disksize.
+  -A|--attach           : Create attached volumes.
+  -b|--bootsize <xxGB>  : Size of boot disk in GB, Default is $bootsize.
+  -d|--disksize <xxGB>  : Size of attached volume(s), Default is $disksize.
+  -D|--disknum   <n>    : Number of additional attached volumes.
   -h|--help             : Display usage and exit.
      --dryrun           : Enable dryrun, no action is taken.
-  -i|--image   <name>   : Set image family as ubuntu (default) or centos.
-  -N|--network <name>   : Define a GCP Network to use for instances.
-  -n|--subnet  <name>   : Used with --network to define the subnet.
+  -i|--image   <name>   : Set image family as 'ubuntu' (default) or 'centos'.
+  -N|--network <name>   : GCP Network name.
+  -n|--subnet  <name>   : GCP Network subnet name. 
   -p|--prefix  <name>   : Prefix name to use for instances.
                           Default prefix is '$prefix'.
   -S|--ssd              : Use SSD as attached disk type.
   -t|--type             : Machine type to use for instances.
                           Default is '$mtype'.
-  -T|--tags <tag1,..>   : List of tags to use for instances.
+  -T|--tags <tag1,..>   : Set of tags to use for instances.
   -x|--use-xfs          : Use the XFS filesystem for attached disks.
   -V|--version          : Show usage info and exit.
-  -z|--zone <name>      : Set GCP zone to use if not gcloud default.
+  -z|--zone <name>      : Set GCP zone to use, default is '$zone'.
   
 Where <action> is 'run'. Any other action enables '--dryrun' 
 followed by a list of names that become '\$prefix-\$name'.
@@ -93,6 +95,7 @@ Will dryrun 3 master nodes: $prefix-m01, $prefix-m02, and $prefix-m03
 # MAIN
 #
 rt=0
+chars=( {b..z} )
 
 while [ $# -gt 0 ]; do
     case "$1" in
@@ -111,6 +114,10 @@ while [ $# -gt 0 ]; do
             disksize="$2"
             shift
             ;;
+        -D|--disknum)
+            disknum=$2
+            shift
+            ;;
         -i|--image)
             imagef="$2"
             shift
@@ -122,12 +129,12 @@ while [ $# -gt 0 ]; do
         --dryrun|--dry-run)
             dryrun=1
             ;;
-        -n|--subnet)
-            subnet="$2"
-            shift
-            ;;
         -N|--network)
             network="$2"
+            shift
+            ;;
+        -n|--subnet)
+            subnet="$2"
             shift
             ;;
         -S|-ssd)
@@ -155,26 +162,26 @@ while [ $# -gt 0 ]; do
         *)
             action="${1,,}"
             shift
-            namelist="$@"
+            names="$@"
             shift $#
             ;;
     esac
     shift
 done
 
-if [ -z "$action" ]; then
+if [[ -z "$action" || -z "$names" ]]; then
     tdh_version
     echo "$usage"
     exit 1
 fi
 
 if [ -n "$network" ] && [ -z "$subnet" ]; then
-    echo "ERROR, Subnet must be provided with --network"
+    echo "Error, Subnet must be provided with --network"
     exit 1
 fi
 
 if [[ ! -e ${tdh_path}/../tools/${format} ]]; then
-    echo "ERROR, cannot locate '$format', is this being run from tdh-gcp root?"
+    echo "Error, cannot locate '$format', is this being run from tdh-gcp root?"
     exit 2
 fi
 
@@ -184,9 +191,9 @@ tdh_version
 if [ "$action" == "run" ] && [ $dryrun -eq 0 ]; then
     dryrun=0
 else
-    echo "Action provided is: '$action'. Use 'run' to execute"
+    printf "$C_CYN -> Action provided is: ${C_NC}'%s'. ${C_CYN}Use${C_NC} 'run' ${C_CYN}to execute. $C_NC \n" $action
     dryrun=1
-    echo "  <DRYRUN> enabled"
+    printf " $C_YEL <DRYRUN> enabled $C_NC \n"
 fi
 
 if [ -n "$zone" ]; then
@@ -194,20 +201,12 @@ if [ -n "$zone" ]; then
     GSCP="$GSCP --zone $zone"
 fi
 
-if [ -n "$namelist" ]; then
-    names="$namelist"
-else
-    echo "Using default of 3 master instances"
-fi
+printf "$C_CYN -> Creating instance(s) ${C_NC}${C_WHT}'%s'${C_NC}${C_CYN}\
+ for ${C_NC}{${C_WHT} ${names} ${C_NC}} \n\n" $mtype
 
-
-echo "Creating master instances '$mtype' for { $names }"
-echo ""
-
-# Create instance
 for name in $names; do
     host="${prefix}-${name}"
-    cmd="${tdh_path}/gcp-compute.sh --prefix $prefix --type $mtype --bootsize $bootsize"
+    cmd="${gcpcompute} --prefix $prefix --type $mtype --bootsize $bootsize"
 
     if [ -n "$imagef" ]; then
         cmd="$cmd --image $imagef"
@@ -216,29 +215,29 @@ for name in $names; do
         cmd="$cmd --network $network --subnet $subnet"
     fi
     if [ -n "$zone" ]; then
-        cmd="$cmd --zone $zone"
+        cmd="$cmd --zone ${zone}"
     fi
     if [ $dryrun -gt 0 ]; then
-        cmd="$cmd --dryrun"
+        cmd="${cmd} --dryrun"
     fi
     if [ $attach -gt 0 ]; then
-        cmd="$cmd --attach --disksize $disksize"
+        cmd="${cmd} --attach --disksize $disksize --disknum $disknum"
     fi
     if [ $ssd -gt 0 ]; then
-        cmd="$cmd --ssd"
+        cmd="${cmd} --ssd"
     fi
     if [ -n "$tags" ]; then
         cmd="$cmd --tags $tags"
     fi
 
-    cmd="$cmd create $name"
+    cmd="${cmd} create ${name}"
     echo "( $cmd )"
 
     ( $cmd )
 
     rt=$?
     if [ $rt -gt 0 ]; then
-        echo "Error in GCP initialization of $host"
+        echo "$PNAME Error in GCP initialization of $host"
         break
     fi
 done
@@ -248,15 +247,13 @@ if [ $rt -gt 0 ]; then
     exit $rt
 fi
 
-
-echo ""
-echo -n " -> Waiting for last host '$host' to respond. . "
+printf "\n${C_CYN} -> Waiting for last host ${C_NC}${C_WHT}'%s'${C_NC}${C_CYN} to respond ${C_NC}. . " $host
 
 if [ $dryrun -eq 0 ]; then
     wait_for_gcphost "$host"
     rt=$?
 else
-    echo "  <DRYRUN skipped>"
+    printf "$C_YEL  <DRYRUN skipped> $C_NC \n"
 fi
 echo ""
 
@@ -269,50 +266,60 @@ fi
 for name in $names; do
     host="${prefix}-${name}"
 
-    echo ""
-    echo " => Bootstrapping host '$host'"
+    printf "\n${C_CYN} -> Bootstrapping host ${C_WHT}'%s' ${C_NC} \n" $host
     #
     # Device format and mount
     if [ $attach -gt 0 ]; then
-        device="/dev/sdb"
-        mountpoint="/data01"
-        cmd="./${format}"
-
-        if [ $xfs -eq 1 ]; then
-            cmd="$cmd --use-xfs"
-        fi
-        cmd="$cmd -f $device $mountpoint"
-
-        echo " -> Formatting and Mount of attached disk"
-        echo "( $GSSH $host --command '$cmd' )"
+        printf "${C_CYN} -> Formatting additional volume(s) ${C_NC}"
         if [ $dryrun -eq 0 ]; then
             ( $GSCP ${tdh_path}/../tools/${format} ${host}: )
             ( $GSSH $host --command "chmod +x $format" )
-            ( $GSSH $host --command "$cmd" )
         fi
 
-        rt=$?
-        if [ $rt -gt 0 ]; then
-            echo "Error in $format for $host"
-            break
+        for (( i=0; i<$disknum; )); do
+            device="/dev/sd${chars[i++]}"
+            volnum=$(printf "%02d" $i)
+            mountpt="/data${volnum}"
+
+            cmd="./${format}"
+
+            if [ $xfs -eq 1 ]; then
+                cmd="$cmd --use-xfs"
+            fi
+            cmd="$cmd -f $device $mountpt"
+
+            echo "( $GSSH $host --command '$cmd' )"
+
+            if [ $dryrun -eq 0 ]; then
+                ( $GSSH $host --command "$cmd" )
+            fi
+
+            rt=$?
+            if [ $rt -gt 0 ]; then
+                echo "Error in $format for $host"
+                break
+            fi
+        done
+    fi
+
+    # prereqs
+    printf "$C_CYN -> Install Prereqs $C_NC \n"
+
+    if [[ $imagef =~ centos ]]; then
+        echo "( $GSSH $host --command 'sudo systemctl stop firewalld; sudo systemctl disable firewalld' )"
+
+        if [ $dryrun -eq 0 ]; then
+            ( $GSSH $host --command "sudo systemctl stop firewalld; sudo systemctl disable firewalld" )
         fi
     fi
 
-    # prereq's
-    # disable  iptables and cups
-    echo " -> Prereqs"
-    echo "( $GSSH $host --command 'sudo systemctl stop firewalld; sudo systemctl disable firewalld' )"
-    if [ $dryrun -eq 0 ]; then
-        ( $GSSH $host --command "sudo systemctl stop firewalld; sudo systemctl disable firewalld" )
-    fi
+    echo "( $GSSH $host --command  sudo ./tdh-prereqs.sh )"
 
-    echo "( $GSSH $host --command sudo ./tdh-prereqs.sh )"
     if [ $dryrun -eq 0 ]; then
         ( $GSCP ${tdh_path}/../etc/bashrc ${host}:.bashrc )
         ( $GSCP ${tdh_path}/../tools/tdh-prereqs.sh ${host}: )
         ( $GSSH $host --command 'chmod +x tdh-prereqs.sh' )
         ( $GSSH $host --command './tdh-prereqs.sh' )
-        ( $GSSH $host --command 'sudo yum install -y ansible ansible-lint' )
     fi
 
     rt=$?
@@ -323,54 +330,37 @@ for name in $names; do
 
     #
     # ssh
-    echo " -> Configure ssh host keys"
+    printf "$C_CYN -> Configure ssh host keys $C_NC \n"
 
-    echo "( $GSSH $host --command \"ssh-keygen -t rsa -b 2048 -N '' -f ~/.ssh/id_rsa; \
+    echo "( $GSSH ${host} --command \"ssh-keygen -t rsa -b 2048 -N '' -f ~/.ssh/id_rsa; \
       cat .ssh/id_rsa.pub >> .ssh/authorized_keys; chmod 600 .ssh/authorized_keys\" )"
 
     if [ $dryrun -eq 0 ]; then
-        ( $GSSH $host --command "ssh-keygen -t rsa -b 2048 -N '' -f ~/.ssh/id_rsa; \
+        ( $GSSH ${host} --command "ssh-keygen -t rsa -b 2048 -N '' -f ~/.ssh/id_rsa; \
           cat .ssh/id_rsa.pub >> .ssh/authorized_keys; chmod 600 .ssh/authorized_keys" )
     fi
 
     if [ -e "$master_id_file" ]; then
         echo "( $GSCP ${master_id_file} ${host}:.ssh/ )"
-        echo "( $GSSH $host --command \"cat .ssh/${master_id} >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys\" )"
+        echo "( $GSSH ${host} --command \"cat .ssh/${master_id} >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys\" )"
 
         if [ $dryrun -eq 0 ]; then
             ( $GSCP ${master_id_file} ${host}:.ssh/ )
-            ( $GSSH $host --command "cat .ssh/${master_id} >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys" )
+            ( $GSSH ${host} --command "cat .ssh/${master_id} >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys" )
         fi
     else
         echo "( $GSCP ${host}:.ssh/id_rsa.pub ${master_id_file} )"
-        echo "( $GSSH $host --command \"cat .ssh/id_rsa.pub >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys\" )"
+        echo "( $GSSH ${host} --command \"cat .ssh/id_rsa.pub >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys\" )"
 
         if [ $dryrun -eq 0 ]; then
-            echo "-> Primary Master Host is '$host'"
+            echo "-> Primary host is '$host'"
             ( $GSCP ${host}:.ssh/id_rsa.pub ${master_id_file} )
             ( $GSSH $host --command "cat .ssh/id_rsa.pub >> .ssh/authorized_keys; chmod 700 .ssh; chmod 600 .ssh/authorized_keys" )
         fi
     fi
 
-    #
-    # push self for ansible
-    cmd="${tdh_path}/${TDH_PUSH} -G"
-    if [ -n "$zone" ]; then
-        cmd="$cmd --zone $zone"
-    fi
-
-    echo "( ${cmd} ${tdh_path}/.. tdh-gcp $host )"
-    if [ $dryrun -eq 0 ]; then
-        ( $cmd ${tdh_path}/.. tdh-gcp $host )
-    fi
-
-    echo "-> Initialization complete for $host"
-    echo ""
+    printf "${C_CYN} -> Initialization complete for ${C_WHT}%s${C_NC} \n" $host
 done
 
-if [ -e "$pwfile" ] && [ $dryrun -eq 0 ]; then
-    ( rm $pwfile )
-fi
-
-echo "$TDH_PNAME finished"
+echo "$TDH_PNAME finished."
 exit $rt
